# Conflicts:
#	acme/primitive/primitive/particle.cpp

+++ acme/operating_system/posix/file.h
@@ -29,7 +29,7 @@ public:
     bool is_opened() const {return m_iFile>=0;}
 
     memsize read(void * p, memsize s);
-    int write(const void * p, memsize s);
+    long write(const void * p, memsize s);
 
 };
